Cambiato nome prodotto e nome produttore che rimangono scritti nel pannello di controllo. Cancello la chiave di registry che memorizza le info di installaizone anche prima di una nuova installazione o di un nuovo aggiornamento per evitare che chiavi "cadaveri" mi rovinino il porcesso di installazione. In aggiornamento rimuovo a mano virtual folder e application pool per evitare che rimangano cadaveri in caso di cambi di nomi tra unq versione e la successiva.

+++ Mago4Butler.BL/BL/RegistryService.cs
@@ -32,7 +32,14 @@ namespace Microarea.Mago4Butler.BL
             {
                 if (key != null)
                 {
-                    key.DeleteSubKeyTree(upgradeCode);
+                    foreach (var subKeyName in key.GetSubKeyNames())
+                    {
+                        if (String.Compare(subKeyName, upgradeCode, StringComparison.InvariantCultureIgnoreCase) == 0)
+                        {
+                            key.DeleteSubKeyTree(upgradeCode);
+                            break;
+                        }
+                    }
                 }
             }
         }
